+++ R/regex_join.R
@@ -8,6 +8,7 @@
 #' @param y A tbl
 #' @param by Columns by which to join the two tables
 #' @param mode One of "inner", "left", "right", "full" "semi", or "anti"
+#' @param ignore_case Whether to be case insensitive (default no)
 #'
 #' @seealso \code{\link{str_detect}}
 #'
@@ -31,9 +32,9 @@
 #'  regex_inner_join(d, by = c(cut = "regex_name"))
 #'
 #' @export
-regex_join <- function(x, y, by = NULL, mode = "inner") {
+regex_join <- function(x, y, by = NULL, mode = "inner", ignore_case = FALSE) {
   match_fun <- function(v1, v2) {
-    stringr::str_detect(v1, v2)
+    stringr::str_detect(v1, stringr::regex(v2, ignore_case = ignore_case))
   }
 
   fuzzy_join(x, y, by = by, match_fun = match_fun, mode = mode)
@@ -42,41 +43,41 @@ regex_join <- function(x, y, by = NULL, mode = "inner") {
 
 #' @rdname regex_join
 #' @export
-regex_inner_join <- function(x, y, by = NULL) {
-  regex_join(x, y, by, mode = "inner")
+regex_inner_join <- function(x, y, by = NULL, ignore_case = FALSE) {
+  regex_join(x, y, by, mode = "inner", ignore_case = ignore_case)
 }
 
 
 #' @rdname regex_join
 #' @export
-regex_left_join <- function(x, y, by = NULL) {
-  regex_join(x, y, by, mode = "left")
+regex_left_join <- function(x, y, by = NULL, ignore_case = FALSE) {
+  regex_join(x, y, by, mode = "left", ignore_case = ignore_case)
 }
 
 
 #' @rdname regex_join
 #' @export
-regex_right_join <- function(x, y, by = NULL) {
-  regex_join(x, y, by, mode = "right")
+regex_right_join <- function(x, y, by = NULL, ignore_case = FALSE) {
+  regex_join(x, y, by, mode = "right", ignore_case = ignore_case)
 }
 
 
 #' @rdname regex_join
 #' @export
-regex_full_join <- function(x, y, by = NULL) {
-  regex_join(x, y, by, mode = "full")
+regex_full_join <- function(x, y, by = NULL, ignore_case = FALSE) {
+  regex_join(x, y, by, mode = "full", ignore_case = ignore_case)
 }
 
 
 #' @rdname regex_join
 #' @export
-regex_semi_join <- function(x, y, by = NULL) {
-  regex_join(x, y, by, mode = "semi")
+regex_semi_join <- function(x, y, by = NULL, ignore_case = FALSE) {
+  regex_join(x, y, by, mode = "semi", ignore_case = ignore_case)
 }
 
 
 #' @rdname regex_join
 #' @export
-regex_anti_join <- function(x, y, by = NULL) {
-  regex_join(x, y, by, mode = "anti")
+regex_anti_join <- function(x, y, by = NULL, ignore_case = FALSE) {
+  regex_join(x, y, by, mode = "anti", ignore_case = ignore_case)
 }