+++ NEWS.md
@@ -1,3 +1,12 @@
+# fuzzyjoin 0.1.3
+
+* Added `interval_join`, which joins tables on cases where (start, end) intervals overlap between the two columns. This adds IRanges from Bioconductor to SUGGESTS.
+* Added `genome_join`, which is a more specific case of `interval_join` that joins tables on based on (chromosome, start, end), where the chromosome must agree and (start, end) must overlap.
+* Added `index_match_fun` argument to `fuzzy_join`, which handles functions (such as `interval_join` and `genome_join`) that operate on the original columns rather than all pairs of columns
+* Fixed bug when matching multiple columns to the same column (#28)
+* Fixed bug in which rows were sometimes duplicated when no distance column was specified (#21)
+* Added more unit tests
+
 # fuzzyjoin 0.1.2
 
 * Fixed bug that failed when single column data frames (not tbl_dfs) were given (#13)